@@ -17,6 +17,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@jest/types": "28.1.3",
     "@types/cookie-parser": "^1.4.3",
     "@types/jsonwebtoken": "^8.5.8",
     "@types/mongodb": "^4.0.7",
@@ -35,27 +36,36 @@
     "express-validator": "^6.14.2",
     "inversify": "^6.0.1",
     "ipware": "^2.0.0",
+    "jest": "28.1.3",
     "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.21",
+    "mongodb-memory-server": "^8.9.0",
     "mongoose": "^6.5.2",
     "mongoose-beautiful-unique-validation": "^7.1.1",
     "nodemailer": "^6.7.7",
     "nodemon": "^2.0.16",
     "reflect-metadata": "^0.1.13",
     "request-ip": "^3.3.0",
+    "supertest": "^6.2.4",
+    "ts-jest": "^28.0.8",
     "ts-node": "^10.8.1",
     "typescript": "^4.7.4",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
+    "@jest-mock/express": "^2.0.1",
     "@types/bcryptjs": "^2.4.2",
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
+    "@types/jest": "^29.0.0",
+    "@types/lodash": "^4.14.184",
     "@types/mongoose": "5.7.29",
     "@types/request-ip": "^0.0.37",
+    "@types/supertest": "^2.0.12",
     "@types/uuid": "^8.3.4",
     "@typescript-eslint/eslint-plugin": "^5.30.0",
     "@typescript-eslint/parser": "^5.30.0",
-    "eslint": "^8.18.0",
+    "eslint": "8.22.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.1.0",
     "prettier": "^2.7.1"
