@@ -3,19 +3,22 @@ import { validationResult } from 'express-validator';
 import { inject, injectable } from 'inversify';
 import requestIp from 'request-ip';
 import { v4 as uuidv4 } from 'uuid';
+import { BlackListTokensRepositoryDB } from '../repositories/black-list-tokens-repository-db';
+import { JwtPassService } from '../utils/jwt-pass-service';
+import { MailService } from '../utils/mail-service';
 import { expiredAccess, expiredRefresh } from '../variables';
 import { AuthRepositoryDB } from '../repositories/auth-repository-db';
-import { blackListTokensRepositoryDB } from '../repositories/black-list-tokens-repository-db';
 import { UsersRepositoryDB } from '../repositories/users-repository-db';
 import { errorFormatter } from '../utils/error-util';
-import { jwtPassService } from '../utils/jwt-pass-service';
-import { mailService } from '../utils/mail-service';
 
 @injectable()
 export class AuthController {
   constructor(
     @inject(UsersRepositoryDB) protected usersRepositoryDB: UsersRepositoryDB,
     @inject(AuthRepositoryDB) protected authRepositoryDB: AuthRepositoryDB,
+    @inject(BlackListTokensRepositoryDB) protected blackListTokensRepositoryDB: BlackListTokensRepositoryDB,
+    @inject(JwtPassService) protected jwtPassService: JwtPassService,
+    @inject(MailService) protected mailService: MailService,
   ) {}
 
   async login(req: express.Request, res: express.Response) {
@@ -25,11 +28,11 @@ export class AuthController {
       return res.status(400).send({ errorsMessages: result.array() });
     }
     if (!isCheck) {
-      return res.send(401);
+      return res.sendStatus(401);
     } else {
       const user = await this.usersRepositoryDB.getUserByLogin(req.body.login);
       if (user) {
-        const refreshToken = jwtPassService.createJwt(user._id, expiredRefresh);
+        const refreshToken = this.jwtPassService.createJwt(user._id, expiredRefresh);
         return res
           .cookie('refreshToken', refreshToken, {
             httpOnly: true,
@@ -38,7 +41,7 @@ export class AuthController {
           .status(200)
           .send(isCheck);
       } else {
-        return res.status(430).send('Db error');
+        return res.status(400).send('Db error');
       }
     }
   }
@@ -51,20 +54,19 @@ export class AuthController {
     }
 
     const confirmationCode = uuidv4();
+    const isSendStatus = await this.mailService.sendEmail(req.body.email, confirmationCode);
 
-    await this.usersRepositoryDB.createUser(
-      req.body.login,
-      req.body.password,
-      req.body.email,
-      userIp!,
-      confirmationCode,
-    );
-    const isSendStatus = await mailService.sendEmail(req.body.email, confirmationCode);
     if (isSendStatus.error) {
-      await this.usersRepositoryDB.deleteUserByLogin(req.body.login);
-      return res.status(400).send('failed delete user');
+      return res.status(400).send('Error with sending email, user was not created');
     } else {
-      return res.status(204).send(isSendStatus.data);
+      await this.usersRepositoryDB.createUser(
+        req.body.login,
+        req.body.password,
+        req.body.email,
+        userIp!,
+        confirmationCode,
+      );
+      return res.sendStatus(204);
     }
   }
 
@@ -75,12 +77,12 @@ export class AuthController {
     }
     const newCode = uuidv4();
     await this.usersRepositoryDB.updateCodeByEmail(req.body.email, newCode);
-    const isSendStatus = await mailService.sendEmail(req.body.email, newCode);
+    const isSendStatus = await this.mailService.sendEmail(req.body.email, newCode);
     if (isSendStatus.error) {
       await this.usersRepositoryDB.deleteUserByLogin(req.body.login);
       return res.status(400).send('failed delete user');
     } else {
-      return res.send(204);
+      return res.sendStatus(204);
     }
   }
 
@@ -89,16 +91,16 @@ export class AuthController {
     if (!result.isEmpty()) {
       return res.status(400).send({ errorsMessages: result.array() });
     } else {
-      return res.send(204);
+      return res.sendStatus(204);
     }
   }
 
   async getRefreshAccessToken(req: express.Request, res: express.Response) {
-    const accessToken = jwtPassService.createJwt(req.user!._id!, expiredAccess);
-    const refreshToken = jwtPassService.createJwt(req.user!._id!, expiredRefresh);
-    const blackToken = await blackListTokensRepositoryDB.addToken(req.cookies.refreshToken);
+    const accessToken = this.jwtPassService.createJwt(req.user!._id!, expiredAccess);
+    const refreshToken = this.jwtPassService.createJwt(req.user!._id!, expiredRefresh);
+    const blackToken = await this.blackListTokensRepositoryDB.addToken(req.cookies.refreshToken);
     if (typeof blackToken === 'string') {
-      res.status(430).send(blackToken);
+      res.status(400).send(blackToken);
     } else {
       return res
         .cookie('refreshToken', refreshToken, {
@@ -112,11 +114,11 @@ export class AuthController {
 
   async logout(req: express.Request, res: express.Response) {
     const confirmedUser = await this.usersRepositoryDB.confirmUserById(req.user!._id!.toString(), false);
-    const blackToken = await blackListTokensRepositoryDB.addToken(req.cookies.refreshToken);
+    const blackToken = await this.blackListTokensRepositoryDB.addToken(req.cookies.refreshToken);
     if (typeof blackToken === 'string' || typeof confirmedUser === 'string') {
-      res.status(430).send(`${blackToken} ${confirmedUser}`);
+      res.status(400).send(`${blackToken} ${confirmedUser}`);
     } else {
-      return res.send(204);
+      return res.sendStatus(204);
     }
   }
 
