@@ -9,32 +9,36 @@ import { authRouter } from './routes/auth-route';
 import { bloggersRouter } from './routes/bloggers-route';
 import { commentsRouter } from './routes/comments-route';
 import { postsRouter } from './routes/posts-route';
+import { quizRouter } from './routes/quiz-route';
 import { testingRouter } from './routes/testing-route';
 import { usersRouter } from './routes/users-route';
 
 config({ path: path.join(__dirname, '..', '.env') });
 
-const app = express();
+export const app = express();
 const port = process.env.PORT || 5000;
 app.use(cors());
 app.use(bodyParser.json());
 app.use(cookieParser());
 
 app.set('trust proxy', true);
-main()
-  .then(() => {
-    app.use('/bloggers', bloggersRouter);
-    app.use('/posts', postsRouter);
-    app.use('/auth', authRouter);
-    app.use('/comments', commentsRouter);
-    app.use('/users', usersRouter);
-    app.use('/testing', testingRouter);
+app.use('/bloggers', bloggersRouter);
+app.use('/posts', postsRouter);
+app.use('/auth', authRouter);
+app.use('/comments', commentsRouter);
+app.use('/users', usersRouter);
+app.use('/testing', testingRouter);
+app.use('/pair-game-quiz', quizRouter);
 
-    app.listen(port, () => {
-      console.log(`Server started at http://localhost:${port}`);
+if (process.env.NODE_ENV !== 'test') {
+  main()
+    .then(() => {
+      app.listen(port, () => {
+        console.log(`Server started at http://localhost:${port}`);
+      });
+    })
+    .catch((error: Error) => {
+      console.error('Database connection failed', error);
+      process.exit();
     });
-  })
-  .catch((error: Error) => {
-    console.error('Database connection failed', error);
-    process.exit();
-  });
+}
