@@ -1,7 +1,9 @@
+import { injectable } from 'inversify';
 import nodemailer from 'nodemailer';
 import { db_service_email_str, db_service_pass_str } from '../connect-db';
 
-export const mailService = {
+@injectable()
+export class MailService {
   async sendEmail(email: string, code: string) {
     const clientPort = '7000';
     const transporter = nodemailer.createTransport({
@@ -14,10 +16,10 @@ export const mailService = {
 
     const clientSrc = `http://localhost:${clientPort}/client-confirm?code=${code}`;
     const mailOptions = {
-      from: '"lismgmk 👻" <lismgmk2@gmail.com>', // sender address
-      to: email, // list of receivers
-      subject: 'Hello', // Subject line
-      html: `<h1>this is a test mail.<a href=${clientSrc}>Confirm here</a></h1>`, // plain text body
+      from: '"lismgmk 👻" <lismgmk2@gmail.com>',
+      to: email,
+      subject: 'Hello',
+      html: `<h1>this is a test mail.<a href=${clientSrc}>Confirm here</a></h1>`,
     };
     let responseEmail: { error: boolean; data: string };
     try {
@@ -29,5 +31,5 @@ export const mailService = {
       responseEmail = { data: error.response, error: true };
     }
     return responseEmail;
-  },
-};
+  }
+}
